pyproject.toml
[project]
name = "Lucy-API"
version = "0.2.0"
description = "Api rest for lucy project"
authors = [
    {name = "Hyromy",email = "formulajoel9@gmail.com"}
]
license = {text = "MIT License"}
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "whitenoise (>=6.12.0,<7.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "requests (>=2.33.1,<3.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B"]

ignore = [
    "E501",
    "B008",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.9,<0.16.0)"
]
testing = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--reuse-db"
filterwarnings = [
    "ignore:No directory at.*staticfiles:UserWarning",
]
